+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `audioMessageBuilder` (no default implementation yet). Thanks @marinkobabic for the PR!
 - Add `videoMessageBuilder` (no default implementation yet)
+- Add custom text matchers to the `TextMessageOptions`. Thanks @jld3103 for the PR!
 - Update dependencies. Requires Dart >= 2.18.0.
 
 ## 1.6.5